Extracted private class and made it public.
